Ticket #370
Test cases(Upload file, Direct supported camps, profile info)
