+++ src/components/ComponentPages/DirectSupportedCamps/DirectSupportedCampsUI/index.tsx
@@ -208,6 +208,7 @@ export default function DirectSupportedCampsUI({
                                   </Link>
                                 </div>
                                 <CloseCircleOutlined
+                                  data-testid="handle_close"
                                   onClick={() => {
                                     handleClose(tag, data.topic_num, data, []),
                                       setValData(tag),
@@ -225,6 +226,7 @@ export default function DirectSupportedCampsUI({
                         {showSaveChanges && idData == data.topic_num ? (
                           <div className={styles.tag_Changes}>
                             <Button
+                              data-testid="save_change_btn"
                               id="saveChangeBtn"
                               className={styles.save_Changes_Btn}
                               onClick={() => {
@@ -237,6 +239,8 @@ export default function DirectSupportedCampsUI({
                               Save Changes
                             </Button>
                             <Button
+                              data-testid="save_change_btn"
+
                               id="revertBtn"
                               className={styles.revert_Btn}
                               onClick={() => {
