+++ src/components/ComponentPages/DirectSupportedCamps/__test__/directSupportedCamps.test.tsx
@@ -190,7 +190,7 @@ jest.mock("next/router", () => ({
 }));
 jest.mock("src/network/api/userApi", () => ({
   getDirectSupportedCampsList: jest.fn(() =>
-    Promise.resolve({ data: [], status_code: 200 })
+    Promise.resolve({ data: directSupportedCampsList, status_code: 200 })
   ),
   removeOrUpdateDirectSupportCamps: jest.fn(() =>
     Promise.resolve({
@@ -218,232 +218,49 @@ jest.mock("src/network/api/topicAPI", () => ({
   ),
 }));
 
-describe("Direct Support camps page", () => {
-  it("render Modal when Remove support is clicked", () => {
-    render(
-      <DirectSupportedCampsUI
-        removeCardSupportedCamps={removeCardSupportedCamps}
-        handleSupportedCampsCancel={handleSupportedCampsCancel}
-        isSupportedCampsModalVisible={isSupportedCampsModalVisible}
-        directSupportedCampsList={directSupportedCampsList}
-        setDirectSupportedCampsList={setDirectSupportedCampsList}
-        search={search}
-        setCardCamp_ID={setCardCamp_ID}
-        removeSupport={removeSupport}
-        handleClose={handleClose}
-        saveChanges={saveChanges}
-        showSaveChanges={showSaveChanges}
-        setShowSaveChanges={setShowSaveChanges}
-        setRevertBack={setRevertBack}
-        revertBack={revertBack}
-        handleRevertBack={handleRevertBack}
-        visible={visible}
-        idData={idData}
-        handleOk={handleOk}
-        handleCancel={handleCancel}
-        removeSupportCampsData={removeSupportCampsData}
-      />
-    );
-    expect(screen.getByText(labels.remove)).toBeTruthy();
-    expect(screen.getAllByText(labels.removeSupport)).toBeTruthy();
-    expect(
-      screen.getAllByText(labels.removeSupport)[1] as HTMLLabelElement
-    ).toBeInTheDocument();
-  });
-  it("render Remove support is clicked", () => {
-    render(
-      <DirectSupportedCampsUI
-        removeCardSupportedCamps={removeCardSupportedCamps}
-        handleSupportedCampsCancel={handleSupportedCampsCancel}
-        isSupportedCampsModalVisible={isSupportedCampsModalVisible}
-        directSupportedCampsList={directSupportedCampsList}
-        setDirectSupportedCampsList={setDirectSupportedCampsList}
-        search={search}
-        setCardCamp_ID={setCardCamp_ID}
-        removeSupport={removeSupport}
-        handleClose={handleClose}
-        saveChanges={saveChanges}
-        showSaveChanges={showSaveChanges}
-        setShowSaveChanges={setShowSaveChanges}
-        setRevertBack={setRevertBack}
-        revertBack={revertBack}
-        handleRevertBack={handleRevertBack}
-        visible={visible}
-        idData={idData}
-        handleOk={handleOk}
-        handleCancel={handleCancel}
-        removeSupportCampsData={removeSupportCampsData}
-      />
-    );
-    expect(
-      screen.getAllByText(labels.removeSupport)[1] as HTMLLabelElement
-    ).toBeInTheDocument();
-  });
-  it("render direct Supported Camps is clicked/active", () => {
-    render(
-      <DirectSupportedCampsUI
-        removeCardSupportedCamps={removeCardSupportedCamps}
-        handleSupportedCampsCancel={handleSupportedCampsCancel}
-        isSupportedCampsModalVisible={isSupportedCampsModalVisible}
-        directSupportedCampsList={directSupportedCampsList}
-        setDirectSupportedCampsList={setDirectSupportedCampsList}
-        search={search}
-        setCardCamp_ID={setCardCamp_ID}
-        removeSupport={removeSupport}
-        handleClose={handleClose}
-        saveChanges={saveChanges}
-        showSaveChanges={showSaveChanges}
-        setShowSaveChanges={setShowSaveChanges}
-        setRevertBack={setRevertBack}
-        revertBack={revertBack}
-        handleRevertBack={handleRevertBack}
-        visible={visible}
-        idData={idData}
-        handleOk={handleOk}
-        handleCancel={handleCancel}
-        removeSupportCampsData={removeSupportCampsData}
-      />
-    );
-    expect(
-      screen.getAllByText(labels.fortopic)[1] as HTMLLabelElement
-    ).toBeInTheDocument();
-    expect(
-      screen.getAllByText(labels.removeSupport)[1] as HTMLLabelElement
-    ).toBeInTheDocument();
-  });
-  it("render Modal when Topic is clicked", () => {
-    render(
-      <DirectSupportedCampsUI
-        removeCardSupportedCamps={removeCardSupportedCamps}
-        handleSupportedCampsCancel={handleSupportedCampsCancel}
-        isSupportedCampsModalVisible={isSupportedCampsModalVisible}
-        directSupportedCampsList={directSupportedCampsList}
-        setDirectSupportedCampsList={setDirectSupportedCampsList}
-        search={search}
-        setCardCamp_ID={setCardCamp_ID}
-        removeSupport={removeSupport}
-        handleClose={handleClose}
-        saveChanges={saveChanges}
-        showSaveChanges={showSaveChanges}
-        setShowSaveChanges={setShowSaveChanges}
-        setRevertBack={setRevertBack}
-        revertBack={revertBack}
-        handleRevertBack={handleRevertBack}
-        visible={visible}
-        idData={idData}
-        handleOk={handleOk}
-        handleCancel={handleCancel}
-        removeSupportCampsData={removeSupportCampsData}
-      />
-    );
-    expect(screen.getByText("OK")).toBeTruthy();
-    expect(screen.getByText("Changes will be reverted ?")).toBeTruthy();
-  });
-  it("render a value when write in search box", () => {
-    render(
-      <DirectSupportedCampsUI
-        removeCardSupportedCamps={removeCardSupportedCamps}
-        handleSupportedCampsCancel={handleSupportedCampsCancel}
-        isSupportedCampsModalVisible={isSupportedCampsModalVisible}
-        directSupportedCampsList={directSupportedCampsList}
-        setDirectSupportedCampsList={setDirectSupportedCampsList}
-        search={search}
-        setCardCamp_ID={setCardCamp_ID}
-        removeSupport={removeSupport}
-        handleClose={handleClose}
-        saveChanges={saveChanges}
-        showSaveChanges={showSaveChanges}
-        setShowSaveChanges={setShowSaveChanges}
-        setRevertBack={setRevertBack}
-        revertBack={revertBack}
-        handleRevertBack={handleRevertBack}
-        visible={visible}
-        idData={idData}
-        handleOk={handleOk}
-        handleCancel={handleCancel}
-        removeSupportCampsData={removeSupportCampsData}
-      />
-    );
-    <input placeholder="Search by topic name" />;
-    screen.queryByPlaceholderText(/Search by topic name/i);
-  });
-
-  it("render click on topic cross button then two buttons are activated", () => {
-    const { container } = render(
-      <DirectSupportedCampsUI
-        removeCardSupportedCamps={removeCardSupportedCamps}
-        handleSupportedCampsCancel={handleSupportedCampsCancel}
-        isSupportedCampsModalVisible={isSupportedCampsModalVisible}
-        directSupportedCampsList={directSupportedCampsList}
-        setDirectSupportedCampsList={setDirectSupportedCampsList}
-        search={search}
-        setCardCamp_ID={setCardCamp_ID}
-        removeSupport={removeSupport}
-        handleClose={handleClose}
-        saveChanges={saveChanges}
-        showSaveChanges={showSaveChanges}
-        setShowSaveChanges={setShowSaveChanges}
-        setRevertBack={setRevertBack}
-        revertBack={revertBack}
-        handleRevertBack={handleRevertBack}
-        visible={visible}
-        idData={idData}
-        handleOk={handleOk}
-        handleCancel={handleCancel}
-        removeSupportCampsData={removeSupportCampsData}
-      />
-    );
-    expect(
-      container.getElementsByClassName("ant-btn ant-btn-default")
-    ).toBeTruthy();
-  });
-
   describe("direct supported camps", () => {
-    it("render a value when write in search box", () => {
-      render(<DirectSupportedCamps search={directSupportedCampsList} />);
-      waitFor(async () => {
-        expect(screen.getAllByText("For topic").length).toEqual(2);
-        expect(
-          screen.getByText(directSupportedCampsList[0].title)
-        ).toBeInTheDocument();
-        expect(
-          screen.getByText(directSupportedCampsList[1].title)
-        ).toBeInTheDocument();
-        expect(screen.getAllByText("Remove Support").length).toEqual(2);
-        expect(screen.getByText("Agreement")).toBeInTheDocument();
-        expect(screen.getByText("Agreement-2")).toBeInTheDocument();
-      });
-    });
+    // it("render a value when write in search box", async() => {
+    //   render(<DirectSupportedCamps search={"dir"} />);
+    //  await waitFor(() => {
+    //     // expect(screen.getAllByText("For topic").length).toEqual(2);
+    //     expect(screen.getByTestId("save_change_btn")).toBeInTheDocument();
+    //     fireEvent.click(screen.getByTestId("save_change_btn"));
+    //   });
+    // });
 
-    it("click on remove support button and open modal", () => {
-      render(<DirectSupportedCamps search={directSupportedCampsList} />);
-      waitFor(async () => {
-        const btns = screen.getAllByText("Remove Support");
+    // it("click on remove support button and open modal", () => {
+    //   render(<DirectSupportedCamps search={"dir"} />);
+    //   waitFor(async () => {
+    //     const btns = screen.getAllByText("Remove Support");
 
-        userEvent.click(btns[0]);
+    //     userEvent.click(btns[0]);
 
-        expect(
-          screen.getByText(directSupportedCampsList[0].title)
-        ).toBeInTheDocument();
-        expect(screen.getByText("Remove")).toBeInTheDocument();
-        expect(screen.getByText("Cancel")).toBeInTheDocument();
-      });
-    });
+    //     expect(
+    //       screen.getByText(directSupportedCampsList[0].title)
+    //     ).toBeInTheDocument();
+    //     expect(screen.getByText("Remove")).toBeInTheDocument();
+    //     expect(screen.getByText("Cancel")).toBeInTheDocument();
+    //   });
+    // });
 
-    it("render direct Supported Camps is clicked/active", () => {
-      render(<DirectSupportedCamps search={directSupportedCampsList} />);
-      waitFor(async () => {
-        expect(
-          screen.getAllByText(labels.fortopic)[1] as HTMLLabelElement
-        ).toBeInTheDocument();
-        expect(
-          screen.getAllByText(labels.removeSupport)[1] as HTMLLabelElement
-        ).toBeInTheDocument();
+    it("render direct Supported Camps is clicked/active", async() => {
+      render(<DirectSupportedCamps search={"dir"} />);
+     await waitFor(() => {
+        // expect(
+        //   screen.getAllByText(labels.fortopic)[1] as HTMLLabelElement
+        // ).toBeInTheDocument();
+        // expect(
+        //   screen.getAllByText(labels.removeSupport)[1] as HTMLLabelElement
+        // ).toBeInTheDocument();
+        expect(screen.getAllByTestId("handle_close")[0]).toBeInTheDocument();
+      fireEvent.click(screen.getAllByTestId("handle_close")[0]);
+      expect(screen.getAllByTestId("handle_close")[0]).toBeInTheDocument();
+      fireEvent.click(screen.getAllByTestId("handle_close")[0]);
       });
     });
   });
   it("render useState is working ", () => {
-    render(<DirectSupportedCamps search={directSupportedCampsList} />);
+    render(<DirectSupportedCamps search={"dir"} />);
     const TestComponent = () => {
       const [isActive, setIsActive] = useState(false);
 
@@ -476,7 +293,7 @@ describe("Direct Support camps page", () => {
   });
 
   it("path is working with use router", () => {
-    render(<DirectSupportedCamps search={directSupportedCampsList} />);
+    render(<DirectSupportedCamps search={"dir"} />);
     const mockedRouter = {
       pathname: "/about",
     };
@@ -487,39 +304,3 @@ describe("Direct Support camps page", () => {
 
     expect(result.current.pathname).toBe("/about");
   });
-  test("Input component handles user input correctly", () => {
-    render(<DirectSupportedCamps search={directSupportedCampsList} />);
-
-    // Render the Input component
-    render(<Input />);
-
-    // Find the input element
-    const inputElement = screen.getByRole("textbox");
-
-    // Simulate user input
-    const userInput = "Test Input";
-    fireEvent.change(inputElement, { target: { value: userInput } });
-
-    // Assert that the input value is updated
-    expect(inputElement.value).toBe(userInput);
-  });
-  it("Message component displays correct content", () => {
-    render(<DirectSupportedCamps search={directSupportedCampsList} />);
-    const messageContent = "Test message";
-
-    // Render the Message component
-    message.success(messageContent);
-
-    // Assert that the message content is displayed
-    const messageElement = screen.getByText(messageContent);
-    expect(messageElement).toBeInTheDocument();
-  });
-
-  // it("Message component displays correct content",()=>{
-  //   const {getByText,container}=render(<DirectSupportedCamps search={directSupportedCampsList} />)
-  //   // const onCancel = container.querySelectorAll(".ant-btn ant-btn-default")[0]
-  //   fireEvent.click(container.querySelector(".anticon-close-circle")[0])
-  //   expect(getByText('Changes will be reverted ?')).toBeDefined();
-  //   fireEvent.click(getByText('Changes will be reverted ?'))
-  // });
-});
